requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
python-decouple==3.8
python-dotenv==1.1.0
